civicpulse-backend/apply_migration.py
from supabase import create_client
import os
from dotenv import load_dotenv

load_dotenv()

SUPABASE_URL = os.getenv("SUPABASE_URL")
SUPABASE_KEY = os.getenv("SUPABASE_SERVICE_ROLE_KEY") # Use service role key for admin tasks if available, else anon might fail for RLS setup
# Fallback to anon key if service key not in env, but RLS setup usually needs admin rights.
if not SUPABASE_KEY:
    SUPABASE_KEY = os.getenv("SUPABASE_ANON_KEY")

if not SUPABASE_URL or not SUPABASE_KEY:
    print("Error: Missing SUPABASE_URL or SUPABASE_SERVICE_ROLE_KEY/ANON_KEY")
    exit(1)

supabase = create_client(SUPABASE_URL, SUPABASE_KEY)

# Read the SQL file
with open("fix_complaint_permissions.sql", "r") as f:
    sql_statements = f.read()

# Execute via RPC if you have a function, but standard client doesn't execute raw SQL easily without a helper.
# Actually, the python client doesn't expose a raw sql method for security.
# But we can try to use a postgres connector if available. 
# Or, if this is Supabase, we might have to ask the User to run it in the SQL Editor.

# WAIT. I cannot reliably run raw SQL via the supabase-js/py client unless I have a stored procedure `exec_sql`.
# I will output the instructions for the user.
print("\n--- ACTION REQUIRED ---")
print("Please copy the content of 'fix_complaint_permissions.sql' and run it in your Supabase SQL Editor.")
print("This will fix the permissions so Officers can update complaint statuses.")
print("-----------------------\n")
