civicpulse-backend/requirements.txt
fastapi
uvicorn
supabase
pydantic
python-dotenv
google-generativeai
openai
